requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==23.0.0
PyYAML==6.0.2
librosa==0.10.2.post1
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
soundfile==0.12.1
tensorflow==2.19.0
tqdm==4.66.5
